Code/requirements.txt
python-gitlab==3.12.0
Flask==2.2.2
Flask-Cors==3.0.10
PyYAML~=6.0